# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transport telemetry API",
    "description": "NDTP, GPS arrival estimates, remote CatBoost inference and dispatcher stream.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/vehicles/{unit_id}": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Vehicle",
        "operationId": "vehicle_api_vehicles__unit_id__get",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicles/{unit_id}/history": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "History",
        "operationId": "history_api_vehicles__unit_id__history_get",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Unit Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 10000,
              "title": "Limit"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Forecast time T, including zone: 2026-09-26T12:30:00Z",
              "title": "Until"
            },
            "description": "Forecast time T, including zone: 2026-09-26T12:30:00Z"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Coordinates": {
        "properties": {
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          },
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lon",
          "lat"
        ],
        "title": "Coordinates"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "History": {
        "properties": {
          "unit_id": {
            "type": "integer",
            "title": "Unit Id"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "retained_count": {
            "type": "integer",
            "title": "Retained Count"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Telemetry"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "capacity",
          "retained_count",
          "count",
          "items"
        ],
        "title": "History"
      },
      "MapMatch": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "matched",
              "ambiguous",
              "unmatched"
            ],
            "title": "Status"
          },
          "snapped_position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "segment_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Id"
          },
          "route_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Id"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "source_packet_id": {
            "type": "string",
            "title": "Source Packet Id"
          },
          "source_event_time": {
            "type": "string",
            "format": "date-time",
            "title": "Source Event Time"
          },
          "graph_version": {
            "type": "string",
            "minLength": 1,
            "title": "Graph Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "source_packet_id",
          "source_event_time",
          "graph_version"
        ],
        "title": "MapMatch",
        "description": "Enrichment attached to Vehicle.position, never a replacement for raw GPS."
      },
      "Prediction": {
        "properties": {
          "sample_id": {
            "type": "string",
            "title": "Sample Id"
          },
          "tr_id": {
            "type": "integer",
            "title": "Tr Id"
          },
          "T": {
            "type": "string",
            "format": "date-time",
            "title": "T"
          },
          "target_stop_id": {
            "type": "string",
            "title": "Target Stop Id"
          },
          "target_time_begin": {
            "type": "string",
            "format": "date-time",
            "title": "Target Time Begin"
          },
          "predicted_delay_s": {
            "type": "number",
            "title": "Predicted Delay S"
          },
          "late_probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Late Probability"
          },
          "risk_alert": {
            "type": "boolean",
            "title": "Risk Alert"
          },
          "early_risk_alert": {
            "type": "boolean",
            "title": "Early Risk Alert"
          },
          "telemetry_stale": {
            "type": "boolean",
            "title": "Telemetry Stale"
          },
          "cur_dev_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cur Dev S"
          },
          "delay_threshold_s": {
            "type": "number",
            "title": "Delay Threshold S"
          },
          "alert_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Alert Threshold"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "explanations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanations"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/PredictionFactor"
            },
            "type": "array",
            "title": "Factors"
          },
          "observations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observations"
          },
          "explanation_method": {
            "type": "string",
            "title": "Explanation Method",
            "default": "unavailable"
          },
          "risk_baseline_log_odds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Baseline Log Odds"
          },
          "risk_other_contribution": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Other Contribution"
          },
          "segment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SegmentContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Recommendations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sample_id",
          "tr_id",
          "T",
          "target_stop_id",
          "target_time_begin",
          "predicted_delay_s",
          "late_probability",
          "risk_alert",
          "early_risk_alert",
          "telemetry_stale",
          "cur_dev_s",
          "delay_threshold_s",
          "alert_threshold",
          "generated_at",
          "model_version"
        ],
        "title": "Prediction"
      },
      "PredictionFactor": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "contribution": {
            "type": "number",
            "title": "Contribution"
          },
          "direction": {
            "type": "string",
            "enum": [
              "increases_risk",
              "decreases_risk"
            ],
            "title": "Direction"
          },
          "method": {
            "type": "string",
            "const": "tree_shap",
            "title": "Method",
            "default": "tree_shap"
          },
          "scale": {
            "type": "string",
            "const": "calibrated_log_odds",
            "title": "Scale",
            "default": "calibrated_log_odds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "feature",
          "label",
          "value",
          "unit",
          "contribution",
          "direction"
        ],
        "title": "PredictionFactor"
      },
      "Recommendation": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Priority"
          },
          "requires_dispatcher": {
            "type": "boolean",
            "title": "Requires Dispatcher",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "title",
          "rationale",
          "priority"
        ],
        "title": "Recommendation"
      },
      "SegmentContext": {
        "properties": {
          "segment_id": {
            "type": "string",
            "title": "Segment Id"
          },
          "graph_version": {
            "type": "string",
            "title": "Graph Version"
          },
          "from_stop_id": {
            "type": "string",
            "title": "From Stop Id"
          },
          "to_stop_id": {
            "type": "string",
            "title": "To Stop Id"
          },
          "from_name": {
            "type": "string",
            "title": "From Name"
          },
          "to_name": {
            "type": "string",
            "title": "To Name"
          },
          "basis": {
            "type": "string",
            "const": "target_approach",
            "title": "Basis",
            "default": "target_approach"
          },
          "geometry_source": {
            "type": "string",
            "title": "Geometry Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "segment_id",
          "graph_version",
          "from_stop_id",
          "to_stop_id",
          "from_name",
          "to_name",
          "geometry_source"
        ],
        "title": "SegmentContext"
      },
      "Telemetry": {
        "properties": {
          "packet_id": {
            "type": "string",
            "title": "Packet Id"
          },
          "tr_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tr Id"
          },
          "unit_id": {
            "type": "integer",
            "title": "Unit Id"
          },
          "event_time": {
            "type": "string",
            "format": "date-time",
            "title": "Event Time"
          },
          "device_event_id": {
            "type": "integer",
            "title": "Device Event Id"
          },
          "location_valid": {
            "type": "boolean",
            "title": "Location Valid"
          },
          "gps_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gps Time"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "alt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          },
          "heading": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "receive_time": {
            "type": "string",
            "format": "date-time",
            "title": "Receive Time"
          },
          "is_hist_data": {
            "type": "boolean",
            "title": "Is Hist Data"
          }
        },
        "type": "object",
        "required": [
          "packet_id",
          "tr_id",
          "unit_id",
          "event_time",
          "device_event_id",
          "location_valid",
          "gps_time",
          "lon",
          "lat",
          "alt",
          "speed",
          "heading",
          "receive_time",
          "is_hist_data"
        ],
        "title": "Telemetry"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Vehicle": {
        "properties": {
          "unit_id": {
            "type": "integer",
            "title": "Unit Id"
          },
          "tr_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tr Id"
          },
          "latest": {
            "$ref": "#/components/schemas/Telemetry"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Telemetry"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Received At"
          },
          "history_count": {
            "type": "integer",
            "title": "History Count"
          },
          "is_stale": {
            "type": "boolean",
            "title": "Is Stale"
          },
          "position_age_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Age S"
          },
          "map_match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MapMatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Prediction"
              },
              {
                "type": "null"
              }
            ]
          },
          "prediction_status": {
            "type": "string",
            "title": "Prediction Status",
            "default": "not_connected"
          },
          "arrival": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Recommendations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "tr_id",
          "latest",
          "position",
          "last_received_at",
          "history_count",
          "is_stale",
          "position_age_s"
        ],
        "title": "Vehicle"
      }
    }
  }
}